+++ api/v1beta1/beskar7cluster_types.go
@@ -78,8 +78,6 @@ type Beskar7ClusterStatus struct {
 // +kubebuilder:object:generate=true
 // +kubebuilder:storageversion
 // +kubebuilder:metadata:labels=cluster.x-k8s.io/v1beta1=v1beta1
-// +kubebuilder:metadata:labels=cluster.x-k8s.io/v1beta2=v1beta1
-
 // Beskar7Cluster is the Schema for the beskar7clusters API.
 type Beskar7Cluster struct {
 	metav1.TypeMeta   `json:",inline"`